pyproject.toml
[project]
name = "pdf-worker"
version = "0.1.0"
description = "CLI tool for resizing/compressing and merging PDFs"
requires-python = ">=3.11"
dependencies = ["pikepdf>=9.0.0", "Pillow>=10.0.0", "PyYAML>=6.0"]

[dependency-groups]
dev = ["pytest>=8.0.0"]

[project.scripts]
pdf-worker = "pdf_worker.cli:main"
ppw = "pdf_worker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_worker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
